Set proper return type of GetSize

+++ lib/size_darwin.go
@@ -3,6 +3,6 @@ package lib
 import "fmt"
 
 // GetSpace returns disk size information for the given drive (bytes).
-func GetSpace(string) (int, int, int, error) {
+func GetSpace(string) (int64, int64, int64, error) {
 	return 0, 0, 0, fmt.Errorf("not implemented")
 }
